Move markdown to markdown.kt + code cleanup
