+++ core/api/core.api
@@ -6122,6 +6122,11 @@ public final class org/jetbrains/kotlinx/dataframe/io/HtmlKt {
 	public static synthetic fun toStaticHtml$default (Lorg/jetbrains/kotlinx/dataframe/DataFrame;Lorg/jetbrains/kotlinx/dataframe/io/DisplayConfiguration;Lorg/jetbrains/kotlinx/dataframe/jupyter/CellRenderer;ZZILjava/lang/Object;)Lorg/jetbrains/kotlinx/dataframe/io/DataFrameHtmlData;
 }
 
+public final class org/jetbrains/kotlinx/dataframe/io/MarkdownKt {
+	public static final fun renderToMarkdown (Lorg/jetbrains/kotlinx/dataframe/DataFrame;Ljava/lang/Integer;Ljava/lang/Integer;ZZZZ)Ljava/lang/String;
+	public static synthetic fun renderToMarkdown$default (Lorg/jetbrains/kotlinx/dataframe/DataFrame;Ljava/lang/Integer;Ljava/lang/Integer;ZZZZILjava/lang/Object;)Ljava/lang/String;
+}
+
 public final class org/jetbrains/kotlinx/dataframe/io/MethodArguments {
 	public static final field Companion Lorg/jetbrains/kotlinx/dataframe/io/MethodArguments$Companion;
 	public fun <init> ()V
